+++ src/Snipcart.php
@@ -89,7 +89,7 @@ class Snipcart extends Plugin
     /**
      * @var string
      */
-    public $schemaVersion = '1.0.7';
+    public $schemaVersion = '1.0.8';
 
 
     // Public Methods