+++ changelog.md
@@ -1,5 +1,16 @@
 # Snipcart Changelog
 
+## 1.3.4 - 2020-06-23
+### Added
+- Added support for using Product Details fields in element queries.
+
+### Changed
+- Improved precision of price, weight, and dimensions by storing as decimals rather than floats.
+- Orders without any shippable items are not sent to ShipStation.
+
+### Fixed
+- Fixed minor styling issues with price field in Craft 3.4.
+
 ## 1.3.3 - 2020-05-18
 ### Changed
 - Exceptions will be thrown if Snipcart’s API is erroring or unresponsive.